fix: circular import crashed the staging app at boot
schemas -> utilities.downloads -> storage.keys triggers storage/__init__ ->
r2.py, which imported content_disposition back out of the partially
initialized downloads module. only main.py's import order (uvicorn's) hits
it, which is why tests and an import check that touched backend.storage
first both passed. content_disposition moves into storage/r2.py — its only
consumer — removing the back-edge; a regression test now imports
backend.main in a fresh interpreter, the way uvicorn does.
