+++ loq.toml
@@ -47,7 +47,7 @@ max_lines = 1266
 
 [[rules]]
 path = "backend/src/backend/storage/r2.py"
-max_lines = 981
+max_lines = 998
 
 [[rules]]
 path = "backend/tests/api/test_audio.py"
